requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.2
django-elasticsearch-dsl==7.3
django-elasticsearch-dsl-drf==0.22.5
django-import-export==3.0.1
django-nine==0.2.7
djangorestframework==3.14.0
elasticsearch==7.17.9
elasticsearch-dsl==7.4.1
et-xmlfile==1.1.0
idna==3.4
MarkupPy==1.14
numpy==1.24.2
odfpy==1.4.1
openpyxl==3.0.10
packaging==23.1
pandas==2.0.0
psycopg2==2.9.6
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tablib==3.2.1
tzdata==2023.3
urllib3==1.26.12
xlrd==2.0.1
xlwt==1.3.0
